+++ src/framework/piral-core/src/components/PiralRoutes.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
-import { Route, Switch, RouteComponentProps, SwitchProps } from 'react-router';
+import { RouteComponentProps, SwitchProps } from 'react-router';
 import { useGlobalState } from '../hooks';
+import { RouteSwitchProps } from '../types';
 
 /**
  * The props used by the PiralRoutes component.
@@ -10,25 +11,23 @@ export interface RoutesProps extends SwitchProps {
    * Sets the component for showing the not found page.
    */
   NotFound: React.ComponentType<RouteComponentProps>;
+  /**
+   * Sets the component for actually switching the routes.
+   */
+  RouteSwitch: React.ComponentType<RouteSwitchProps>;
 }
 
 /**
  * The component for defining the exclusive routes to be used.
  */
-export const PiralRoutes: React.FC<RoutesProps> = ({ NotFound, ...props }) => {
+export const PiralRoutes: React.FC<RoutesProps> = ({ NotFound, RouteSwitch, ...props }) => {
   const routes = useGlobalState((s) => s.routes);
   const pages = useGlobalState((s) => s.registry.pages);
+  const paths = [];
 
-  return (
-    <Switch {...props}>
-      {Object.keys(routes).map((url) => (
-        <Route exact key={url} path={url} component={routes[url]} />
-      ))}
-      {Object.keys(pages).map((url) => (
-        <Route exact key={url} path={url} component={pages[url].component} />
-      ))}
-      <Route component={NotFound} />
-    </Switch>
-  );
+  Object.keys(routes).map((path) => paths.push({ path, Component: routes[path] }));
+  Object.keys(pages).map((path) => paths.push({ path, Component: pages[path].component }));
+
+  return <RouteSwitch NotFound={NotFound} paths={paths} {...props} />;
 };
 PiralRoutes.displayName = 'Routes';