+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed issue placing *tsconfig.json* in empty template
 - Fixed issue placing *piral-layout.jsx* in empty template
+- Fixed page not found message when doing pilet development
 
 ## 0.14.13 (February 27, 2022)